Fix hang on cross-compiled system (#7)
When deploying the cross-compiled build to a Hetzner arm64 machine, it hangs. The only reference to a similar issue I've been able to find is [here](https://users.rust-lang.org/t/cross-compiled-hello-world-hangs-on-aarch64/56321/3).

The fix was to unset `PKG_CONFIG_SYSROOT_DIR`, as suggested [here](https://felix-knorr.net/posts/2023-01-11-cross-compiling-rust.html).